Find in Files: Report regexp errors in the same way as in the Editor
